% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework3.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient Descent}
\usage{
gradient_descent(fx, gradf, x0, t = NULL, max_iter = 100, tol = 0.001)
}
\arguments{
\item{fx}{handle to function that returns objective function values}

\item{gradf}{handle to function that returns gradient of objective function}

\item{x0}{initial parameter estimate}

\item{t}{step-size}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}
}
\description{
Gradient Descent
}
