% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework1.R
\name{plot_restrictions}
\alias{plot_restrictions}
\title{Plot restrictions}
\usage{
plot_restrictions(fx, rx, nRow = 3, nCol = 3)
}
\arguments{
\item{fx}{handle to function in question}

\item{rx}{handle to function that returns a sequence of feasible sequence t and x + tv}

\item{nRow}{number of row plots}

\item{nCol}{number of column plots}
}
\description{
Plot restrictions
}
